+++ Assets/Scripts/Enemy/BeetleWave.cs
@@ -6,8 +6,8 @@ public class BeetleWave : MonoBehaviour
     private ObjectPooler destroyEffectPool;
     private AudioSource destroySound;
     private AudioSource hitSound;
-    [SerializeField] private int lives;
-    [SerializeField] private int maxLives;
+    [SerializeField] private float lives;
+    [SerializeField] private float maxLives;
     [SerializeField] int expToGive;
     [SerializeField] int livesMultipler = 1;
     private FlashWhite flashWhite;
@@ -50,7 +50,7 @@ public class BeetleWave : MonoBehaviour
     {
         
     }
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         lives -= damage;
         if(lives > 0)
