+++ Assets/Scripts/Meteor.cs
@@ -12,9 +12,9 @@ public class Meteor : MonoBehaviour
 
     [SerializeField] private Sprite[] sprites;
 
-    [SerializeField] int lives;
-    [SerializeField] private int maxLives;
-    [SerializeField] private int dmg = 1;
+    [SerializeField] float lives;
+    [SerializeField] private float maxLives;
+    [SerializeField] private float dmg = 1;
 
     private ObjectPooler destroyEffectPool;
 
@@ -61,7 +61,7 @@ public class Meteor : MonoBehaviour
     }
     
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         Debug.Log("Meteor taking damage");
         flashWhite.Flash();
